Ejercicios/ejercicio2-Biblioteca/src/main/java/ar/edu/unlp/info/oo2/biblioteca/Biblioteca.java
package ar.edu.unlp.info.oo2.biblioteca;

import java.util.ArrayList;
import java.util.List;

public class Biblioteca {
	private List<Socio> socios;
	private JSONExporter exporter;

	public Biblioteca() {
		socios = new ArrayList<>();
		exporter= new VoorheesExporter(); 
	}

	public void agregarSocio(Socio socio) {
		socios.add(socio);
	}

	/**
	 * Retorna la representación JSON de la colección de socios.
	 */
	public String exportarSocios() {
		return exporter.exportar(socios);
	}

	public JSONExporter getExporter() {
		return exporter;
	}

	public void setExporter(JSONExporter exporter) {
		this.exporter = exporter;
	}
}
